Implemented jump and checks
Jump code block is implemented and works accordingly. The interpreter class has checks to make sure blocks are valid, and stops the the runner if not. Destination color blocks are now on the Back ground array instead of level array for discrete functionality of the level array. Today is a good day.